에러처리 추가

+++ src/codes/codes.controller.ts
@@ -20,22 +20,34 @@ export class CodesController {
   constructor(private codesService: CodesService) {}
   @Post()
   async createCode(@Res() res: Response, @Body() createCodeDto: CreateCodeDto) {
-    await this.codesService.createCodes(createCodeDto);
-    res.status(HttpStatus.CREATED).send('생성 완료');
+    try {
+      await this.codesService.createCodes(createCodeDto);
+      res.status(HttpStatus.CREATED).send('생성 완료');
+    } catch (e) {
+      res.status(500).send(e.message);
+    }
   }
 
   // 부모(최상위까지) 코드 가져오기
   @Get('/parents/:code')
   async getParentCode(@Res() res: Response, @Param('code') code: string) {
-    const parentsCodes = await this.codesService.getParentsCodesInfo(code);
-    res.status(HttpStatus.OK).send(parentsCodes);
+    try {
+      const parentsCodes = await this.codesService.getParentsCodesInfo(code);
+      res.status(HttpStatus.OK).send(parentsCodes);
+    } catch (e) {
+      res.status(500).send(e.message);
+    }
   }
 
   // 자식 코드 가져오기
   @Get('/child/:code')
   async getChildCode(@Res() res: Response, @Param('code') code: string) {
-    const childCode = await this.codesService.getChildsCodesInfo(code);
-    res.status(HttpStatus.OK).send(childCode);
+    try {
+      const childCode = await this.codesService.getChildsCodesInfo(code);
+      res.status(HttpStatus.OK).send(childCode);
+    } catch (e) {
+      res.status(500).send(e.message);
+    }
   }
 
   @Patch(':id')
@@ -44,13 +56,27 @@ export class CodesController {
     @Param('id', ParseIntPipe) id: number,
     @Body() updateCodeData: UpdateCodeDto,
   ) {
-    await this.codesService.update(id, updateCodeData);
-    res.status(HttpStatus.OK).send('업데이트 완료');
+    try {
+      const codeInfo = await this.codesService.getMyCodeInfoByCodeOrId({ id });
+      if (!codeInfo) return res.status(500).send('No Update Data');
+
+      await this.codesService.update(id, updateCodeData);
+      res.status(HttpStatus.OK).send('업데이트 완료');
+    } catch (e) {
+      res.status(500).send(e.message);
+    }
   }
 
   @Delete(':id')
   async remove(@Res() res: Response, @Param('id', ParseIntPipe) id: number) {
-    await this.codesService.deleteCode(id);
-    res.status(HttpStatus.OK).send('삭제 완료');
+    try {
+      const codeInfo = await this.codesService.getMyCodeInfoByCodeOrId({ id });
+      if (!codeInfo) return res.status(500).send('No Delete Data');
+
+      await this.codesService.deleteCode(id);
+      res.status(HttpStatus.OK).send('삭제 완료');
+    } catch (e) {
+      res.status(500).send(e.message);
+    }
   }
 }
